chore: revert rename DownloadWalletZip p2
